Compatibility with open bluedragon 3.0
Added a patch for compatibility with open bluedragon 3.0

The recurse and filter options for cfdirectory work differently in
openbd. Added a compatibility component which uses a custom method to
emulate the same behaviour for cfdirectoy in openbluedragon.

Changed the keys for META_ABSTRACT and META_GENERIC as openbluedragon
throws an error when a structure key name contains a colon.

The patch has been tested with openbd 3.0 and Railo 4